remove button from view and added a edit button.
