+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/noteInput/NoteLayout.kt
@@ -20,7 +20,7 @@ import com.example.habittrackerapp.model.Note
 fun NoteLayout() {
     val navController = LocalNavController.current
     Column() {
-        Button(onClick = { navController.navigate("NoteListScreenRoute") },
+        Button(onClick = { navController.navigate(Routes.ViewList.route) },
             modifier = Modifier
                 .fillMaxWidth()
                 .padding(16.dp,0.dp,16.dp,2.dp)){
