+++ HabitTrackerApp/app/src/main/java/com/example/habittrackerapp/noteInput/screens/SingleNoteElementDisplay.kt
@@ -1,0 +1,118 @@
+package com.example.habittrackerapp.noteInput.screens
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonColors
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import coil.compose.AsyncImage
+import com.example.habittrackerapp.R
+import com.example.habittrackerapp.LocalNavController
+import com.example.habittrackerapp.model.NotesViewModel
+import com.example.habittrackerapp.model.NotesViewModelFactory
+
+
+@Composable
+fun SingleNoteElementDisplay(id: String,
+                             notesViewModel: NotesViewModel = viewModel(factory= NotesViewModelFactory())
+) {
+    val navController = LocalNavController.current
+    val note by notesViewModel.currentNote.collectAsState()
+    notesViewModel.getNote(id)
+    Column{
+        Row(
+            modifier = Modifier
+                .align(Alignment.CenterHorizontally)
+        )
+        {
+            Button(onClick = { navController.navigate(Routes.ViewList.route)},
+                modifier = Modifier
+                    .padding(top=10.dp),
+                colors =  ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.tertiary)) {
+                Text("View List",  modifier = Modifier.padding(start=30.dp,end=30.dp));
+            }
+
+            Button(
+                onClick = { navController.navigate("EditNoteScreenRoute/${note.id}") },
+                modifier = Modifier
+                    .padding(start=10.dp,top=10.dp),
+                colors =  ButtonDefaults.buttonColors(containerColor = MaterialTheme.colorScheme.tertiary)
+            ) {
+                Text(text = "edit",  modifier = Modifier
+                    .padding(start=40.dp,end=40.dp))
+            }
+        }
+
+        Card(
+            modifier = Modifier
+                .fillMaxWidth()
+                .height(600.dp)
+                .padding(16.dp, top = 2.dp, 16.dp, 16.dp),
+            shape = RoundedCornerShape(16.dp),
+        ) {
+            Column(
+                modifier = Modifier
+                    .padding(20.dp)
+                    .verticalScroll(rememberScrollState())
+            )
+            {
+
+                if (note.id == id) {
+                    LoadImage(url = note.urlImage!!)
+
+                    //title text field
+                    Text(
+                        text = note.title,
+                        modifier = Modifier
+                            .padding(20.dp, 0.dp)
+                            .fillMaxWidth()
+                    )
+                    //body/content description text field
+                    Text(
+                        text = note.description,
+                        modifier = Modifier
+                            .height(250.dp)
+                            .padding(20.dp, 0.dp),
+                    )
+                    //add remove button and edit at the top??
+
+                }
+            }
+
+        }
+    }
+}
+
+@Composable
+fun LoadImage(url:String) {
+    AsyncImage(
+        model = url,
+        error= painterResource(id = R.drawable.no_image_placeholder),
+        contentDescription = "user image",
+        modifier = Modifier.padding(20.dp) )
+}
+
+
